auth-service/requirements.txt
flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
flask-cors==4.0.0
pyjwt==2.8.0
python-jose==3.3.0
requests==2.31.0